added blueprints
